Fix more things holy cow
