+++ packages/create-toolpad-app/src/templates/nextjs/nextjs-pages/employeeRoute.ts
@@ -1,24 +1,24 @@
-const route = `import { NextRequest, NextResponse } from 'next/server';
-import { getEmployeesStore, setEmployeesStore } from '../../../../employeesStore';
-import type { Employee } from '../../../../data/employees';
+const route = `import { NextApiRequest, NextApiResponse } from 'next';
+import { getEmployeesStore, setEmployeesStore } from '../../../employeesStore';
+import type { Employee } from '../../../data/employees';
 import type { OmitId } from '@toolpad/core/Crud';
 
-export async function GET(req: NextRequest, { params }: { params: { id: string } }) {
-  const { id: employeeId } = params;
+export async function getEmployee(req: NextApiRequest, res: NextApiResponse) {
+  const { id: employeeId } = req.query;
 
   const employeesStore = getEmployeesStore();
 
   const employeeToShow = employeesStore.find((employee) => employee.id === Number(employeeId));
 
   if (!employeeToShow) {
-    return NextResponse.json({ error: 'Employee not found' }, { status: 404 });
+    return res.status(404).json({ error: 'Employee not found' });
   }
-  return NextResponse.json(employeeToShow);
+  return res.status(200).json(employeeToShow);
 }
 
-export async function PUT(req: NextRequest, { params }: { params: { id: string } }) {
-  const body: Partial<OmitId<Employee>> = await req.json();
-  const { id: employeeId } = params;
+export async function updateEmployee(req: NextApiRequest, res: NextApiResponse) {
+  const body: Partial<OmitId<Employee>> = req.body;
+  const { id: employeeId } = req.query;
 
   const employeesStore = getEmployeesStore();
 
@@ -35,19 +35,35 @@ export async function PUT(req: NextRequest, { params }: { params: { id: string }
   );
 
   if (!updatedEmployee) {
-    return NextResponse.json({ error: 'Employee not found' }, { status: 404 });
+    return res.status(404).json({ error: 'Employee not found' });
   }
-  return NextResponse.json(updatedEmployee);
+  return res.status(200).json(updatedEmployee);
 }
 
-export async function DELETE(req: NextRequest, { params }: { params: { id: string } }) {
-  const { id: employeeId } = params;
+export async function deleteEmployee(req: NextApiRequest, res: NextApiResponse) {
+  const { id: employeeId } = req.query;
 
   const employeesStore = getEmployeesStore();
 
   setEmployeesStore(employeesStore.filter((employee) => employee.id !== Number(employeeId)));
 
-  return NextResponse.json({ success: true });
+  return res.status(200).json({ success: true });
+}
+
+export default function handler(req: NextApiRequest, res: NextApiResponse) {
+  if (req.method === 'GET') {
+    return getEmployee(req, res);
+  }
+
+  if (req.method === 'PUT') {
+    return updateEmployee(req, res);
+  }
+
+  if (req.method === 'DELETE') {
+    return deleteEmployee(req, res);
+  }
+
+  return res.status(405).json({ message: 'Method Not Allowed' });
 }
 `;
